+++ .mypy.ini
@@ -53,7 +53,7 @@ modules =
     azul.indexer.action_controller,
     azul.indexer.mirror_controller,
     azul.service.app_controller,
-    azul.indexer.mirror_service,
+    azul.indexer.mirror_file_service,
     azul.digests,
     lambdas.indexer.app,
     azul.schemas,
