+++ test/indexer/test_mirror_controller.py
@@ -30,7 +30,7 @@ from azul.indexer import (
 from azul.indexer.mirror_controller import (
     MirrorController,
 )
-from azul.indexer.mirror_service import (
+from azul.indexer.mirror_file_service import (
     MirrorService,
 )
 from azul.json import (
